@@ -1,25 +1,25 @@
 module github.com/markusressel/fan2go
 
-go 1.18
+go 1.21
 
 require (
 	github.com/asecurityteam/rolling v2.0.4+incompatible
 	github.com/guptarohit/asciigraph v0.5.6
 	github.com/labstack/echo-contrib v0.15.0
-	github.com/labstack/echo/v4 v4.11.1
+	github.com/labstack/echo/v4 v4.11.4
 	github.com/looplab/tarjan v0.1.0
 	github.com/md14454/gosensors v0.0.0-20180726083412-bded752ab001
 	github.com/mgutz/ansi v0.0.0-20200706080929-d51e80ef957d
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/oklog/run v1.1.0
-	github.com/prometheus/client_golang v1.16.0
-	github.com/pterm/pterm v0.12.67
-	github.com/spf13/cobra v1.7.0
-	github.com/spf13/viper v1.16.0
+	github.com/prometheus/client_golang v1.18.0
+	github.com/pterm/pterm v0.12.75
+	github.com/spf13/cobra v1.8.0
+	github.com/spf13/viper v1.18.2
 	github.com/stretchr/testify v1.8.4
 	github.com/tomlazar/table v0.1.2
-	go.etcd.io/bbolt v1.3.7
-	golang.org/x/exp v0.0.0-20230905200255-921286631fa9
+	go.etcd.io/bbolt v1.3.8
+	golang.org/x/exp v0.0.0-20240119083558-1b970713d09a
 )
 
 require (
